Extract VisionDevice to abstract class
